feat(client): create a new home view and change the previous home view to be cardsets view
