+++ client/src/router/index.ts
@@ -10,6 +10,11 @@ const router = createRouter({
       component: HomeView,
     },
     {
+      path: '/cardsets',
+      name: 'cardsets',
+      component: () => import('../views/CardsetsView.vue'),
+    },
+    {
       path: '/login',
       name: 'login',
       component: () => import('../views/LoginView.vue'),
